Move SegmentExtractingNoder to noding package
Make SegmentExtractingNoder public

JTS' commits:
* dr-jts/jts@d45d5d7d9db4a25ba776c877a2e39848fe9fc7b4
* dr-jts/jts@5cce15848b8e82015cb2ae90c146687ed9e68483
