+++ src/NetTopologySuite/Operation/OverlayNG/CoverageUnion.cs
@@ -1,4 +1,5 @@
 ﻿using NetTopologySuite.Geometries;
+using NetTopologySuite.Noding;
 
 namespace NetTopologySuite.Operation.OverlayNg
 {
